Split test_list_actors.py to avoid TIMEOUT
