+++ python/ray/tests/BUILD
@@ -144,6 +144,9 @@ py_test_module_list(
     "test_unhandled_error.py",
     "test_top_level_api.py",
     "test_list_actors.py",
+    "test_list_actors_2.py",
+    "test_list_actors_3.py",
+    "test_list_actors_4.py",
     "test_usage_stats.py",
   ],
   size = "small",
@@ -332,6 +335,9 @@ py_test_module_list(
     "test_asyncio.py",
     "test_multiprocessing.py",
     "test_list_actors.py",
+    "test_list_actors_2.py",
+    "test_list_actors_3.py",
+    "test_list_actors_4.py",
   ],
   size = "medium",
   extra_srcs = SRCS,
